ai_dev_team/requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
gradio>=5.0.0
python-dotenv>=1.0.0
duckduckgo-search>=7.0.0
